==> 2.  Visualizations.Rmd <==
```{r}
if (!require("pacman")) install.packages("pacman")
pacman::p_load(tidyverse, magrittr, scales, here, sf)
```

# Global growth accounting

## (a) Per capita GDP growth rates by region

```{r}
regions <- read_csv("regions_growth.csv")
```

```{r}
library(scales)

regions %>%
  mutate(pcap = y/pop,
         pcap_growth = ((pcap - lag(pcap) ) / pcap) * 100) %>% 
  filter(year >= 1985) %>%
  ggplot(aes(year, pcap_growth, 
             color = fct_reorder2(region, pcap_growth, desc(id)))) +
  geom_line(size = .5) +
  labs(title = "Crecimiento económico en el mundo",
       subtitle = "Tasas de crecimiento del ingreso per cápita",
       caption = "Datos: Penn World Table (v 10.0)") +
  scale_y_continuous(labels = comma, breaks = scales::pretty_breaks(n = 10)) +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 10)) +
  guides(color = guide_legend(nrow = 1)) +
  theme(
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    plot.title = element_text(
      face = "bold",
      hjust = .5,
      margin = margin(10, 0, 5, 0)
    ),
    plot.subtitle = element_text(hjust = .5, 
                                 margin = margin(0, 0, 15, 0)),
    plot.caption = element_text(color = "gray50", 
                                size = 8, 
                                margin = margin(15, 0, 0, 0)),
    legend.position = "bottom",
    legend.title = element_blank(),
    panel.background = element_rect(fill = "gray95")
  )
```

```{r}
library(here)
ggsave(filename = here("Plots","Plot1.png"))
```

## (b) Solow decomposition function

```{r}
solow <- function(data, y, l, k, labsh) {
  data %>%
  mutate(
    # Logs of output, capital and labor
    y = log({{y}}),
    k = log({{k}}),
    l = log({{l}}),
    # First differences (growth rates)
    dy = ({{y}} - lag({{y}})) * 100,
    dk = ({{k}} - lag({{k}})) * 100,
    dl = ({{l}} - lag({{l}})) * 100,
    # Share of capital in the economy
    capsh = 1 - {{labsh}},
    # Contributions of factors in the output growth
    k_cont = dk * capsh,
    l_cont = dl * {{labsh}},
    tfp_cont = dy - k_cont - l_cont
  ) %>%
  na.omit()
}
```

## (c) Growth accounting for Guatemala

```{r}
gtm <- read_csv("guatemala_growth.csv")
```

```{r}
gtm %>%
  solow(y = y,
        l = l,
        k = k,
        labsh = labsh) %>%
  select(country, year, k_cont, l_cont, tfp_cont) %>%
  pivot_longer(-c(country, year)) %>%
  ggplot(aes(year, value, fill = name)) +
  geom_col() +
  labs(title = "Contabilidad del crecimiento para Guatemala",
       subtitle = "Contribución de los factores productivos (1955 – 2019)",
       caption = "Datos: Penn World Table (v 10.0)") +
  guides(color = guide_legend(nrow = 1)) +
  scale_fill_discrete(name = "Contribución:",
                      labels = c("Capital", "Trabajo", "PTF")) +
  scale_y_continuous(labels = comma, breaks = scales::pretty_breaks(n = 10)) +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 15)) +
  guides(color = guide_legend(nrow = 1)) +
  theme(
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    plot.title = element_text(
      face = "bold",
      hjust = .5,
      margin = margin(10, 0, 5, 0)
    ),
    plot.subtitle = element_text(hjust = .5,
                                 margin = margin(0, 0, 15, 0)),
    plot.caption = element_text(
      color = "gray50",
      size = 8,
      margin = margin(15, 0, 0, 0)
    ),
    legend.position = "bottom",
    legend.title = element_blank(),
    panel.background = element_rect(fill = "gray95")
  )
```

```{r}
ggsave(filename = here("Plots","Plot2.png"))
```

## (d) Growth accounting by region

```{r}
regions %>%
  group_by(region) %>%
  filter(year %in% c(1984:2019)) %>%
  solow(y = y,
        l = l,
        k = k,
        labsh = labsh) %>%
  select(region, k_cont, l_cont, tfp_cont) %>%
  summarise_if(is.numeric, mean) %>%
  select(region, k_cont, l_cont, tfp_cont) %>%
  pivot_longer(-c(region)) %>%
  ggplot(aes(region, value, fill = name)) +
  geom_col() +
  labs(title = "Contabilidad del crecimiento a nivel mundial",
       subtitle = "Contribución de los factores productivos (1985 – 2019)",
       caption = "Datos: Penn World Table (v 10.0)") +
  guides(color = guide_legend(nrow = 1)) +
  scale_fill_discrete(name = "Contribución:",
                      labels = c("Capital", "Trabajo", "PTF")) +
  scale_y_continuous(labels = comma, breaks = scales::pretty_breaks(n = 5)) +
  guides(color = guide_legend(nrow = 1)) +
  theme(
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    plot.title = element_text(
      face = "bold",
      hjust = .5,
      margin = margin(10, 0, 5, 0)
    ),
    plot.subtitle = element_text(hjust = .5,
                                 margin = margin(0, 0, 15, 0)),
    plot.caption = element_text(
      color = "gray50",
      size = 8,
      margin = margin(15, 0, 0, 0)
    ),
    legend.position = "bottom",
    legend.title = element_blank(),
    panel.background = element_rect(fill = "gray95")
  )
```

```{r}
ggsave(filename = here("Plots","Plot3.png"))
```

## (e) Growth accounting by country

```{r}
countries <- read_csv("countries_growth.csv")
```

```{r}
countries %>%
  group_by(country) %>%
  filter(year %in% c(1984:2019)) %>%
  solow(y = y,
        l = l,
        k = k,
        labsh = labsh) %>%
  select(country, dy, k_cont, l_cont, tfp_cont) %>%
  summarise_if(is.numeric, mean) %>%
  pivot_longer(-c(country, dy)) %>% 
  arrange(desc(dy)) %>% 
  mutate(country = as.factor(country),
         country = fct_reorder(country, desc(dy))) %>%
  ggplot(aes(country, value, fill = name)) +
  geom_col() +
  labs(title = "Contabilidad del crecimiento por país",
       subtitle = "Contribución de los factores productivos (1985 – 2019)",
       caption = "Datos: Penn World Table (v 10.0)") +
  guides(color = guide_legend(nrow = 1)) +
  scale_fill_discrete(name = "Contribución:",
                      labels = c("Capital", "Trabajo", "PTF")) +
  scale_y_continuous(labels = comma, breaks = scales::pretty_breaks(n = 10)) +
  guides(color = guide_legend(nrow = 1)) +
  theme(
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    axis.text.x = element_text(angle = 90, hjust = 1, vjust = .4),
    plot.title = element_text(
      face = "bold",
      hjust = .5,
      margin = margin(10, 0, 5, 0)
    ),
    plot.subtitle = element_text(hjust = .5,
                                 margin = margin(0, 0, 15, 0)),
    plot.caption = element_text(
      color = "gray50",
      size = 8,
      margin = margin(15, 0, 0, 0)
    ),
    legend.position = "bottom",
    legend.title = element_blank(),
    panel.background = element_rect(fill = "gray95")
  )
```

```{r}
ggsave(filename = here("Plots","Plot4.png"), width = 15, height = 8)
```

## (f) World map

```{r}
countries <- read_csv("countries_growth.csv")
```

```{r}
# Region coordinates
library(sf)
mapdata <- map_data("world") %>% as_tibble()

# Group, solow decomposition and summarise
countries %<>%
  group_by(country) %>%
  filter(year %in% c(1984:2019)) %>%
  solow(y = y,
        l = l,
        k = k,
        labsh = labsh) %>%
  select(region = country, dy, k_cont, l_cont, tfp_cont) %>%
  summarise_if(is.numeric, mean)

# If else for capital, labour and TFP
countries %<>%
  mutate(fct = if_else(
              (k_cont > l_cont & k_cont > tfp_cont), "Capital",
              if_else((l_cont > k_cont & l_cont > tfp_cont), "Trabajo", "PTF")),
         fct = as.factor(fct),
         fct = fct_relevel(fct, "Capital", "Trabajo", "PTF")
  )
```

```{r}
# Which countries don't match?
anti_join(countries, mapdata)

# How are they written in the map dataframe?
mapdata %>% 
  group_by(region) %>% 
  distinct(region) %>% 
  filter(region == "")

# Recode 'em in the countries tibble
countries %<>% 
  mutate(region = as.factor(region),
         region = fct_recode(region,
    "Bolivia" = "Bolivia (Plurinational State of)",
    "Cape Verde" = "Cabo Verde",
    "Ivory Coast" = "Côte d'Ivoire",
    "Swaziland" = "Eswatini",
    "Iran" = "Iran (Islamic Republic of)",
    "Laos" = "Lao People's DR",
    "South Korea" = "Republic of Korea",
    "Trinidad" = "Trinidad and Tobago",
    "Tanzania" = "U.R. of Tanzania: Mainland",
    "UK" = "United Kingdom",
    "USA" = "United States",
    "Venezuela" = "Venezuela (Bolivarian Republic of)"
  ))

# Join dataframes
mapdata <- left_join(mapdata, countries, by = "region")

# Change NA's from fct column 
mapdata %<>% mutate(fct = fct_explicit_na(fct, na_level = "Sin datos"))
```

```{r}
mapdata %>%
  ggplot(aes(x = long, y = lat, group = group)) +
  geom_polygon(aes(fill = fct), color = "white", size = .1) +
  coord_cartesian(ylim = c(-55, 85)) +
  guides(color = guide_legend(nrow = 1)) +
  labs(title = "Factor productivo más relevante en el crecimiento económico",
       subtitle = "desde 1985 a 2019",
       caption = "Datos: Penn World Table (v 10.0)") +
  scale_fill_manual(values = c("#F8766D", "#00BA38", "#619CFF", "gray50")) +
  scale_y_continuous() +
  guides(color = guide_legend(nrow = 1)) +
  theme(
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    axis.ticks.x = element_blank(),
    axis.text.x = element_blank(),
    axis.ticks.y = element_blank(),
    axis.text.y = element_blank(),
    plot.title = element_text(
      face = "bold",
      hjust = .5,
      margin = margin(10, 0, 5, 0)
    ),
    plot.subtitle = element_text(hjust = .5,
                                 margin = margin(0, 0, 15, 0)),
    plot.caption = element_text(
      color = "gray50",
      size = 8,
      margin = margin(15, 0, 0, 0)
    ),
    legend.position = "bottom",
    legend.title = element_blank(),
    panel.background = element_rect(fill = "gray95"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank()
  )
```

```{r}
ggsave(filename = here("Plots","Plot5.png"))
```

## (g) Extreme poverty

```{r}
poverty <- read_csv("regions_poverty.csv")
```

```{r}
poverty %>% 
  filter(year >= 1985) %>%
  ggplot(aes(year, pov, 
             color = fct_reorder2(region, pov, desc(id)))) +
  geom_line(size = .5) +
  labs(title = "Porcentaje de la población viviendo en extrema pobreza",
       subtitle = "(1.90 dólares por día) desde 1985 a 2019",
       caption = "Datos: Banco Mundial") +
  scale_y_continuous(labels = comma, breaks = scales::pretty_breaks(n = 10)) +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 10)) +
  guides(color = guide_legend(nrow = 1)) +
  theme(
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    plot.title = element_text(
      face = "bold",
      hjust = .5,
      margin = margin(10, 0, 5, 0)
    ),
    plot.subtitle = element_text(hjust = .5, 
                                 margin = margin(0, 0, 15, 0)),
    plot.caption = element_text(color = "gray50", 
                                size = 8, 
                                margin = margin(15, 0, 0, 0)),
    legend.position = "bottom",
    legend.title = element_blank(),
    panel.background = element_rect(fill = "gray95")
  )
```

```{r}
ggsave(filename = here("Plots","Plot6.png"))
```
