==> 3. Econometrics.Rmd <==
```{r}
if (!require("pacman")) install.packages("pacman")
pacman::p_load(tidyverse, magrittr)
```

# Contribución de los factores productivos al crecimiento económico y su relación con la pobreza extrema: Evidencia para 91 economías con datos de panel para el periodo de 1985 a 2019

## (a) Importing data frames

```{r}
growth <- read_csv("countries_growth.csv")
poverty <- read_csv("countries_poverty.csv")
data <- growth %>% left_join(poverty)
```

## (b) Solow decomposition

```{r}
solow <- function(data, y, l, k, labsh) {
  data %>%
  mutate(
    # Logs of output, capital and labor
    y = log({{y}}),
    k = log({{k}}),
    l = log({{l}}),
    # First differences (growth rates)
    dy = ({{y}} - lag({{y}})) * 100,
    dk = ({{k}} - lag({{k}})) * 100,
    dl = ({{l}} - lag({{l}})) * 100,
    # Share of capital in the economy
    capsh = 1 - {{labsh}},
    # Contributions of factors in the output growth
    k_cont = dk * capsh,
    l_cont = dl * {{labsh}},
    tfp_cont = dy - k_cont - l_cont
  ) %>%
  na.omit()
}

data_solow <- 
  data %>%
  group_by(country) %>%
  solow(y = y,
        l = l,
        k = k,
        labsh = labsh) %>% 
  select(country:year, dy, k_cont, l_cont, tfp_cont, pov) %>% 
  ungroup()
```

## (d) Exploratory data analysis

```{r}
library(plotly)

lines <- function(data, x, y, color) {
  p <- 
    data %>%
    ggplot(aes({{x}}, {{y}}, color = {{color}})) +
    geom_line() 
  ggplotly(p)
}

data_solow %>% 
  group_by(country) %>% 
  lines(year, pov, country)
```

## (e) Econometrics

### Simple regression models

```{r}
library(lmtest)
library(gridExtra)

# Simple regression models
m1 <- data_solow %>% lm(pov ~ dy, data = . ) 
m2 <- data_solow %>% lm(pov ~ k_cont + l_cont + tfp_cont, data = . )

# Plots
plotlist <- function(data, x, y) {
  require(ggplot2)
  require(purrr)
  y <- enquo(y)

  map(x, function(x)
    ggplot(data, aes(!!sym(x), !!y)) +
    geom_point() +
    geom_smooth(method = "lm", formula = y ~ x)
  )
}

x <- c("k_cont","l_cont","tfp_cont")
plotlist(m1, x, pov) %>% grid.arrange(grobs = .)
plotlist(m2, x, pov) %>% grid.arrange(grobs = .)

# Summary
m1 %>% summary()
m2 %>% summary()
```

### «Before and after» comparisons

```{r}
# Subset
data85 <- data_solow %>% filter(year == 1985)
data19 <- data_solow %>% filter(year == 2019)

# Compute the differences
diff_pov <- data19$pov - data85$pov
diff_k <- data19$k_cont - data85$k_cont
diff_l <- data19$l_cont - data85$l_cont
diff_tfp <- data19$tfp_cont - data85$tfp_cont

# Regression with differenced data
m3 <- lm(diff_pov ~ diff_k + diff_l + diff_tfp)

# Plot
x <- c("diff_k","diff_l","diff_tfp")
plotlist(m3, x, diff_pov) %>% grid.arrange(grobs = .)

# Summary
m3 %>% summary()
```

### Fixed effects regression

```{r}
pacman::p_load(panelr, plm, broom)

# Declare data frame as panel data
panel  <-  panel_data(data_solow, id = country, wave = year)
```

```{r}
# Fixed effects
m4 <- plm(pov ~ dy,
          data = panel,
          index = c("country", "year"),
          model = "within")

m5 <- plm(pov ~ k_cont + l_cont + tfp_cont,
          data = panel,
          index = c("country", "year"),
          model = "within")

m4 %>% summary()
```

### Time fixed effects

```{r}
# Time fixed effects
m6 <- plm(pov ~ dy,
          data = panel,
          index = c("country", "year"),
          model = "within",
          effect = "twoways")

m7 <- plm(pov ~ k_cont + l_cont + tfp_cont,
          data = panel,
          index = c("country", "year"),
          model = "within",
          effect = "twoways")

# Summary
m5 %>% summary()
```

```{r}
library(stargazer)

# Gather clustered standard errors in a list
# (Adjustment for autocorrelation + heteroskedasticity)
rob_se <- list(sqrt(diag(vcovHC(m1, type = "HC1"))),
               sqrt(diag(vcovHC(m2, type = "HC1"))),
               sqrt(diag(vcovHC(m3, type = "HC1"))),
               sqrt(diag(vcovHC(m4, type = "HC1"))),
               sqrt(diag(vcovHC(m5, type = "HC1"))),
               sqrt(diag(vcovHC(m6, type = "HC1"))),
               sqrt(diag(vcovHC(m7, type = "HC1"))))

stargazer(m1, m2, m3, m4, m5, m6, m7,
          digits = 3,
          header = FALSE,
          type = "text", 
          se = rob_se,
          title = "Test",
          model.numbers = FALSE,
          column.labels = c("(1)", "(2)", "(3)", "(4)", "(5)", "(6)", "(7)"))
```

```{r}
# Export stargazer
stargazer(m1, m2, m3, m4, m5, m6, m7,
          digits = 3,
          header = FALSE,
          type = "html", 
          se = rob_se,
          title = "Test",
          model.numbers = FALSE,
          column.labels = c("(1)", "(2)", "(3)", "(4)", "(5)", "(6)", "(7)"),
          out = "Resultados.htm")
```
